move consts to lint/init
