+++ __init__.py
@@ -1,4 +1,2 @@
-"""Provides an API version and shared constants."""
+"""Provides an API version for plugins."""
 VERSION = 4
-WARNING = "warning"
-ERROR = "error"
